pyproject.toml
[project]
name = "ai-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sshtunnel>=0.4.0",
    "uvicorn>=0.38.0",
]
